Data in Game class.
