+++ app/src/main/java/es/ulpgc/tfm/ecocsgo/model/Equipment.kt
@@ -1,11 +1,10 @@
 package es.ulpgc.tfm.ecocsgo.model
 
-class Equipment {
-    var referenceKey: String = ""
-    var cost: Int = 0
+abstract class Equipment {
+    //var referenceKey: String = ""
+    var cost: String = ""
     var name: String = ""
-    var numeration: HashMap<HashMap<String, Int>, Int>? = null
-    var reward: Int = 0
-    var team: Int = 0
-    var acceptedCategories: Int = 0
+    var numeration: HashMap<String, HashMap<String, String>>? = null
+    var team: String = ""
+    //var acceptedCategories: String = ""
 }
