+++ app/src/main/java/es/ulpgc/tfm/ecocsgo/model/Grenade.kt
@@ -1,0 +1,4 @@
+package es.ulpgc.tfm.ecocsgo.model
+
+class Grenade : Equipment() {
+}
