+++ app/src/main/java/es/ulpgc/tfm/ecocsgo/model/Weapon.kt
@@ -1,0 +1,5 @@
+package es.ulpgc.tfm.ecocsgo.model
+
+class Weapon : Equipment() {
+    var reward: Int = 0
+}
